StringCompress: swap "model" and "hprof file"
The layout model field can be of fixed size (up to 70 chars) while file path can be much longer.
So it's better to swap them and show the file path as a last column.

On the 92 line it was missing %s in pattern so the mode not even printed.
On the 269 line I had to add double quotes into value so it will be properly padded.

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>